added nav_editor frontend class
